+++ twoWithOne/OneChoiceServer.h
@@ -6,17 +6,17 @@
 class OneChoiceServer {
 private:
     OneChoiceStorage* storage;
-    void getAESRandomValue(unsigned char* keyword, int cnt, unsigned char* result);
-    int numberOfBins, sizeOfEachBin;
+    void getAESRandomValue(unsigned char* keyword, long cnt, unsigned char* result);
+    long numberOfBins, sizeOfEachBin;
     bool profile = false;
 
 public:
-    OneChoiceServer(int dataIndex, bool inMemory, bool overwrite, bool profile);
-    void clear(int index);
+    OneChoiceServer(long dataIndex, bool inMemory, bool overwrite, bool profile);
+    void clear(long index);
     virtual ~OneChoiceServer();
-    void storeCiphers(int dataIndex, vector<vector<pair<prf_type, prf_type> > > ciphers);
-    vector<prf_type> search(int dataIndex, prf_type token, int  keywordCnt);
-    vector<pair<prf_type, prf_type> > getAllData(int dataIndex);
+    void storeCiphers(long dataIndex, vector<vector<pair<prf_type, prf_type> > > ciphers);
+    vector<prf_type> search(long dataIndex, prf_type token, long  keywordCnt);
+    vector<pair<prf_type, prf_type> > getAllData(long dataIndex);
 
 };
