+++ twoWithOne/Storage.h
@@ -27,19 +27,19 @@ private:
     vector< unordered_map<prf_type, prf_type, PRFHasher> > data;
     //stxxl::unordered_map<prf_type, prf_type, PRFHasher, CompareLess, SUB_BLOCK_SIZE, SUB_BLOCKS_PER_BLOCK>** diskData;
     vector<string> filenames;
-    int dataIndex;
+    long dataIndex;
     prf_type nullKey;
     string fileAddressPrefix;
 
 public:
-    int seekgCount = 0;
-    int KEY_VALUE_SIZE = (2 * AES_KEY_SIZE + sizeof (int));
-    Storage(bool inMemory, int dataIndex, string fileAddressPrefix, bool profile);
+    long seekgCount = 0;
+    long KEY_VALUE_SIZE = (2 * AES_KEY_SIZE + sizeof (long));
+    Storage(bool inMemory, long dataIndex, string fileAddressPrefix, bool profile);
     bool setup(bool overwrite);
-    void insert(int dataIndex, map<prf_type, prf_type> ciphers);
-    vector<prf_type> getAllData(int dataIndex);
-    void clear(int index);
-    prf_type find(int index, prf_type mapKey, bool& found);
+    void insert(long dataIndex, map<prf_type, prf_type> ciphers);
+    vector<prf_type> getAllData(long dataIndex);
+    void clear(long index);
+    prf_type find(long index, prf_type mapKey, bool& found);
     virtual ~Storage();
